% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFKnockoffs.R
\name{MFKnockoffs.threshold}
\alias{MFKnockoffs.threshold}
\title{Threshold for the knockoff filter}
\usage{
MFKnockoffs.threshold(W, q = 0.1, method = c("knockoff+", "knockoff"))
}
\arguments{
\item{W}{the test statistics}

\item{q}{target FDR (false discovery rate)}

\item{method}{either 'knockoff or 'knockoff+'}
}
\value{
The threshold for variable selection.
}
\description{
Computes the threshold for the knockoff filter.
}
