% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MFKnockoffs.R
\name{print.MFKnockoffs.result}
\alias{print.MFKnockoffs.result}
\title{Print results for the knockoff filter}
\usage{
\method{print}{MFKnockoffs.result}(x, ...)
}
\arguments{
\item{x}{the output of a call to MFKnockoffs.filter}

\item{...}{unused}
}
\description{
Prints the list of variables selected by the knockoff filter and the corresponding function call.
}
