% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_sdp.R
\name{MFKnockoffs.knocks.vectorize_matrix}
\alias{MFKnockoffs.knocks.vectorize_matrix}
\title{Vectorize a matrix into the SCS format}
\usage{
MFKnockoffs.knocks.vectorize_matrix(M)
}
\description{
Vectorize a matrix into the SCS format
}
\keyword{internal}
